Move hardware modules into their own package
